Create categories file to outsource array
