+++ client/src/components/Categories.js
@@ -1,0 +1,54 @@
+import light from '../themes/light.js';
+
+const categories = [
+  {
+    icon: 'food',
+    value: 'food',
+    fill: light.colors.action,
+    title: 'Essen'
+  },
+  {
+    icon: 'transport',
+    value: 'transport',
+    fill: light.colors.action,
+    title: 'Transport'
+  },
+  {
+    icon: 'airplane',
+    value: 'airplane',
+    fill: light.colors.action,
+    title: 'Flüge'
+  },
+  {
+    icon: 'hotel',
+    value: 'hotel',
+    fill: light.colors.action,
+    title: 'Hotel'
+  },
+  {
+    icon: 'drinks',
+    value: 'drinks',
+    fill: light.colors.action,
+    title: 'Party'
+  },
+  {
+    icon: 'shopping',
+    value: 'shopping',
+    fill: light.colors.action,
+    title: 'Einkauf'
+  },
+  {
+    icon: 'action',
+    value: 'action',
+    fill: light.colors.action,
+    title: 'Freizeit'
+  },
+  {
+    icon: 'smiley',
+    value: 'smiley',
+    fill: light.colors.action,
+    title: 'Sonstiges'
+  }
+];
+
+export default categories;
